refactor: reorganize /upload file
